fix: merge regex testing and order of preParser
